Update problem-1.py
python

+++ problem-1.py
@@ -1,3 +1,7 @@
+# Problem-1: Create a small calculator which performs operations such as Addition, Subtraction, Multiplication and Division using class.
+#   Calculator inputs :> ‘a’, ‘b’ and ‘type of operation’
+#   Datatype :> ‘a’ = double, ‘b’ = double, ‘type of operation’ = string
+
 class calculator:
     def __init__(self,a,b,type):
         self.val1=a
@@ -15,8 +19,8 @@ class calculator:
                 return self.val1-self.val2
             case _:
                 return  "You Entered Wrong operation"
-a=int(input('Enter first number : '))
-b=int(input('Enter another number : '))
+a=float(input('Enter first number : '))
+b=float(input('Enter another number : '))
 operation=input('Enter the Operation (div,mul,add,sub) : ')
 obj=calculator(a,b,operation)
 print("Result : ",obj.calculate())
